Extract tooltip positioning logic into hook
